--- front-end/src/components/footer/footer.js
@@ -1,38 +1,0 @@
-import React from "react";
-import { Typography } from "@material-ui/core";
-
-function Copyright() {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {"Copyright © "}
-      <Link color="inherit" href="https://material-ui.com/">
-        Your Website
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
-
-class footer extends React.Component {
-  render() {
-    return (
-      <footer className={classes.footer}>
-        <Typography variant="h6" align="center" gutterBottom>
-          Footer
-        </Typography>
-        <Typography
-          variant="subtitle1"
-          align="center"
-          color="textSecondary"
-          component="p"
-        >
-          Have a nice cooking time with this website!
-        </Typography>
-        <Copyright />
-      </footer>
-    );
-  }
-}
-
-export default footer;
